Update Delete Category feature

+++ src/com/bookstore/service/CategoryServices.java
@@ -11,6 +11,7 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import com.bookstore.dao.BookDAO;
 import com.bookstore.dao.CategoryDAO;
 import com.bookstore.entity.Category;
 
@@ -96,9 +97,18 @@ public class CategoryServices {
 
 	public void deleteCategory() throws ServletException, IOException {
 		int categoryId = Integer.parseInt(request.getParameter("id"));
-		categoryDAO.delete(categoryId);
+		BookDAO bookDAO = new BookDAO();
+		long numberOfBooks = bookDAO.countByCategory(categoryId);
+		String message;
 		
-		String message = "The category with ID " + categoryId + " has been removed successfully.";
+		if (numberOfBooks > 0) {
+			message = "Could not delete the category (ID: %d) because it currently contains some books.";
+			message = String.format(message, categoryId);
+		} else {
+			categoryDAO.delete(categoryId);
+			message = "The category with ID " + categoryId + " has been removed successfully.";
+		}
+		
 		listCategory(message);
 	}
